Add database view for search history
All the data we want for a search history is available on the searches
table. It's not clear where the logic for retrieving that data belongs,
though - since the data we want isn't represented by actual records in
that table, a scope seems not quite right, because a scope should return
a selection of records from a table based on query params.

I went with a view because a view behaves like a table in the database,
so we can interact with it using regular ActiveRecord methods.